adjust ability and add ghost sword ability
