+++ common/src/main/java/com/iafenvoy/iceandfire/entity/EntityGhostSword.java
@@ -1,6 +1,7 @@
 package com.iafenvoy.iceandfire.entity;
 
 import com.google.common.collect.Lists;
+import com.iafenvoy.iceandfire.IceAndFire;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
@@ -29,18 +30,18 @@ public class EntityGhostSword extends PersistentProjectileEntity {
     private List<Entity> hitEntities;
     private int knockbackStrength;
 
-    public EntityGhostSword(EntityType<? extends PersistentProjectileEntity> type, World worldIn) {
+    public EntityGhostSword(EntityType<? extends EntityGhostSword> type, World worldIn) {
         super(type, worldIn);
         this.setDamage(9F);
     }
 
-    public EntityGhostSword(EntityType<? extends PersistentProjectileEntity> type, World worldIn, double x, double y, double z, float r, float g, float b) {
+    public EntityGhostSword(EntityType<? extends EntityGhostSword> type, World worldIn, double x, double y, double z, float r, float g, float b) {
         this(type, worldIn);
         this.setPosition(x, y, z);
         this.setDamage(9F);
     }
 
-    public EntityGhostSword(EntityType<? extends PersistentProjectileEntity> type, World worldIn, LivingEntity shooter, double dmg, ItemStack from) {
+    public EntityGhostSword(EntityType<? extends EntityGhostSword> type, World worldIn, LivingEntity shooter, double dmg, ItemStack from) {
         super(type, shooter, worldIn, ItemStack.EMPTY, from);
         this.setDamage(dmg);
     }
