+++ common/src/main/java/com/iafenvoy/iceandfire/item/ability/Ability.java
@@ -6,6 +6,7 @@ import net.minecraft.text.Text;
 import java.util.List;
 
 public interface Ability {
-    void activeAbility(LivingEntity target, LivingEntity attacker);
-    void addAbilityDescription(List<Text> tooltip);
+    boolean isEnable();
+    void active(LivingEntity target, LivingEntity attacker);
+    void addDescription(List<Text> tooltip);
 }
